legion/requirements/base.txt
dill==0.2.7.1
python-interface==1.4.0
pandas==0.22.0
docker==3.1.3
python-consul==0.7.2
flask==0.12.2
numpy==1.13.3
Pillow==4.3.0
Jinja2==2.10
statsd==3.2.2
kubernetes==5.0.0
PyYAML==3.12
pip==9.0.3
tabulate==0.8.2